+++ contrib/scripts/check-crates.sh
@@ -31,7 +31,6 @@ buildargs=(
     "-p nostr-keyring --features async"
     "-p nostr-sdk"                                                # No default features
     "-p nostr-sdk --features all-nips"                            # Only NIPs features
-    "-p nostr-sdk --features tor"                                 # Embedded tor client
     "-p nostr-sdk --all-features"                                 # All features
     "-p nostr-relay-builder"
     "-p nostr-connect"