+++ crates/nwc/CHANGELOG.md
@@ -34,6 +34,7 @@
 ### Changed
 
 - Rename `NWC` to `NostrWalletConnect`
+- Bump MSRV to 1.85.0 (https://github.com/rust-nostr/nostr/pull/1267)
 
 ### Added